+++ src/collision/Grid.js
@@ -35,6 +35,22 @@ var Common = require('../core/Common');
     };
 
     /**
+     * The width of a single grid bucket.
+     *
+     * @property bucketWidth
+     * @type number
+     * @default 48
+     */
+
+    /**
+     * The height of a single grid bucket.
+     *
+     * @property bucketHeight
+     * @type number
+     * @default 48
+     */
+
+    /**
      * Updates the grid.
      * @method update
      * @param {grid} grid
@@ -62,8 +78,8 @@ var Common = require('../core/Common');
                 continue;
 
             // don't update out of world bodies
-            if (body.bounds.max.x < 0 || body.bounds.min.x > world.bounds.width
-                || body.bounds.max.y < 0 || body.bounds.min.y > world.bounds.height)
+            if (body.bounds.max.x < world.bounds.min.x || body.bounds.min.x > world.bounds.max.x
+                || body.bounds.max.y < world.bounds.min.y || body.bounds.min.y > world.bounds.max.y)
                 continue;
 
             var newRegion = _getRegion(grid, body);
@@ -198,7 +214,7 @@ var Common = require('../core/Common');
      * @return {string} bucket id
      */
     var _getBucketId = function(column, row) {
-        return column + ',' + row;
+        return 'C' + column + 'R' + row;
     };
 
     /**